requirements.txt
Flask==2.2.5
librosa==0.10.0
numpy==1.24.3
scipy==1.10.1
dtw-python==1.2.3
praat-parselmouth==0.4.4
soundfile==0.12.1